Update README with license info for pngs
Seems like even these simple screenshots need a license according to an eclass discussion post so, just in case. Licenses are unfortunately outside of my comfort zone so, I'm not sure if I'm also supposed to put my name beside "Abram Hindle" on line 13 of the readme in this commit too. I'll just leave it alone. It should be fine right? My name's on there so it's fine right!?